summary

+++ Company/company_list.php
@@ -5,29 +5,12 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Company List</title>
  <link rel="stylesheet" href="../Sidebar/styles.css">
-    
- <style>
-            body{
-                font-family: system-ui, -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell, 'Open Sans', 'Helvetica Neue', sans-serif;
-                color: #202557;
-            }
-            table {
-                width: 80%;
-                border-collapse: collapse;
-                background-color: #F5F5F5;
-            }
-            th, td {
-                padding: 10px;
-                border: 1px solid lightgray;
-                text-align: left;
-            }
-
-</style>
+    <link rel="stylesheet" href="../CSS/bootstrap_css.css">
 </head>
 <body id="body-pd" class="body-pd" >
 <?php
 include_once("../Sidebar/sidebar.html");
-?>    
+?>
 <h2>Company List</h2>
 <div >
     <table class='table table-bordered table-sm'>
@@ -74,7 +57,7 @@ include_once("../Sidebar/sidebar.html");
                         <td class="p-2">'.htmlspecialchars($data['URL']).'</td>
                         <td class="p-2">
                             <a href="summary_company.php?company_id='.htmlspecialchars($data['company_id']).'" class="btn btn-info btn-sm">View</a>
-                            <a href="delete_company.php?company_id='.htmlspecialchars($data['company_id']).'" class="btn btn-danger btn-sm">Delete</a>
+                            <a href="delete_company.php?company_id='.htmlspecialchars($data['company_id']).'" class="btn btn-danger btn-sm" onclick="return confirmDelete()">Delete</a>
                         </td>
                     </tr>';
                 }
@@ -90,6 +73,10 @@ include_once("../Sidebar/sidebar.html");
 </div>
 
 <script src="../Sidebar/main.js"></script>
-<script src="https://unpkg.com/ionicons@5.1.2/dist/ionicons.js"></script>
+<script>
+    function confirmDelete() {
+        return confirm("Are you sure you want to delete this customer? All related data will be permanently removed.");
+    }
+</script>
 </body>
 </html>
